+++ 42WebGPU-native/engine/sound/IAudioOutput.h
@@ -8,7 +8,7 @@ class IAudioOutput {
 public:
     virtual ~IAudioOutput() = default;
     virtual bool init() = 0;
-    virtual void enqueueData(const std::vector<uint8_t>& pcmData) = 0;
+    virtual void enqueueData(const std::vector<float>& pcmData) = 0;
     virtual void flush() = 0;
     virtual void setVolume(float volume) = 0;
     virtual float getVolume() = 0;
