+++ 42WebGPU-native/engine/video/VideoDecoder.h
@@ -65,7 +65,7 @@ private:
     void open(const std::string& filename, std::unique_ptr<IVideoDecoder> videoDecoder, std::unique_ptr<IAudioOutput> audioOutput);
 
     bool decodeVideoFrame();
-    bool decodeAudioFrame(std::vector<uint8_t>& outPcmData);
+    bool decodeAudioFrame(std::vector<float>& outPcmData);
    
     AVBufferRef* m_hwDeviceContext = nullptr;
     AVDictionary* options = nullptr;
